Put recipe links in unordered list and add shrimp alfredo link to index
Putting multiple links in a row without a list would cause them all to
fall on the same line in the browser. This commit puts the recipe links
in index.html in an unordered list to prevent this from happening.
A link to the shrimp alfredo recipe web page was also added to the
index.