Add images directory along with 3 jpg files
Prior to this, I had forgotten to commit the entire images directory
and the jpg files inside of it to git. This commit will add the images
directory and the images currently inside of it (fried-chicken.jpg,
shrimp-alfredo.jpg, and tiramisu.jpg) that were used in the commits
prior to this one. I will do my best to not make this mistake in the
future.